requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
beanie>=1.26.0
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
websockets>=13.0
